+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-No changes yet
+### Changed
+
+* Swap TextBlock for RichTextBlock in Picto and Split components (#225)
 
 ## [1.1.0]